Moves caching helpers from utils to cacheutils
The aim is to keep the utils package separated by
different topics.
